requirements.txt
#
####### requirements.txt #######
#
click
requests
PyYAML
